+++ src/main/java/site/scalarstudios/scalarpower/gui/ScalarPowerMenus.java
@@ -7,6 +7,7 @@ import site.scalarstudios.scalarpower.machines.generator.coal.CoalGeneratorMenu;
 import site.scalarstudios.scalarpower.machines.generator.barometric.BarometricGeneratorMenu;
 import site.scalarstudios.scalarpower.machines.generator.culinary.CulinaryGeneratorMenu;
 import site.scalarstudios.scalarpower.machines.generator.entropy.EntropyGeneratorMenu;
+import site.scalarstudios.scalarpower.machines.generator.geothermal.GeothermalGeneratorMenu;
 import site.scalarstudios.scalarpower.machines.grinder.DoubleGrinderMenu;
 import site.scalarstudios.scalarpower.machines.grinder.GrinderMenu;
 import site.scalarstudios.scalarpower.machines.macerator.DoubleMaceratorMenu;
@@ -31,6 +32,9 @@ public final class ScalarPowerMenus {
 
     public static final DeferredHolder<MenuType<?>, MenuType<CulinaryGeneratorMenu>> CULINARY_GENERATOR_MENU = MENUS
             .register("culinary_generator", () -> IMenuTypeExtension.create(CulinaryGeneratorMenu::new));
+
+    public static final DeferredHolder<MenuType<?>, MenuType<GeothermalGeneratorMenu>> GEOTHERMAL_GENERATOR_MENU = MENUS
+            .register("geothermal_generator", () -> IMenuTypeExtension.create(GeothermalGeneratorMenu::new));
 
     public static final DeferredHolder<MenuType<?>, MenuType<BarometricGeneratorMenu>> BAROMETRIC_GENERATOR_MENU = MENUS
             .register("barometric_generator", () -> IMenuTypeExtension.create(BarometricGeneratorMenu::new));