+++ src/main/java/site/scalarstudios/scalarpower/item/ScalarPowerCreativeTabs.java
@@ -26,6 +26,7 @@ public class ScalarPowerCreativeTabs {
             /* Generators */
             event.accept(ScalarPowerBlocks.COAL_GENERATOR.get());
             event.accept(ScalarPowerBlocks.CULINARY_GENERATOR.get());
+            event.accept(ScalarPowerBlocks.GEOTHERMAL_GENERATOR.get());
             event.accept(ScalarPowerBlocks.BAROMETRIC_GENERATOR.get());
             event.accept(ScalarPowerBlocks.ENTROPY_GENERATOR.get());